pyproject.toml
[project]
name = "taxpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.86",
    "ete3>=3.1.3",
    "ipython>=9.7.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
]
